\name{examine}
\alias{examine}
\title{
Examine one or more variables
}
\description{
This function is one of many R functions enabling users to assess variable descriptives. It has been developed to mimic SPSS' 'EXAMINE' syntax command as closely as possible to ease the transition for new R users and facilitate teaching courses where both programs are taught alongside each other.
}
\usage{
examine(..., stem = TRUE, plots = TRUE, extremeValues = 5,
        qqCI = TRUE)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{\dots}{
The first argument is a list of variables to provide descriptives for. Because these are the first arguments, the other arguments must be named explicitly so R does not confuse them for something that should be part of the dots.
}
  \item{stem}{
Whether to display a stem and leaf plot.
}
  \item{plots}{
Whether to display the plots generated by the \code{\link{dataShape}} function.
}
  \item{extremeValues}{
How many extreme values to show at either end (the highest and lowest values). When set to FALSE, no extreme values are shown.
}
  \item{qqCI}{
Whether to display confidence intervals in the QQ-plot.
}
}
\details{
This function basically just calls the \code{\link{descr}} function, optionally supplemented with calls to \code{\link{stem}}, \code{\link{dataShape}}.
}
\value{
A list that is displayed when printed.
}
\author{
Gjalt-Jorn Peters

Maintainer: Gjalt-Jorn Peters <gjalt-jorn@userfriendlyscience.com>
}

\seealso{
\code{\link{descr}}, \code{\link{dataShape}}, \code{\link{stem}}
}
\examples{
examine(mtcars$mpg, stem=FALSE, plots=FALSE);

}
\keyword{ univar }
